Talantix/requirements.txt
sentence-transformers
faiss-cpu
nltk
requests
torch